tests/wpt/tests/html/semantics/document-metadata/the-style-element/non-parser-created-doc.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8">
    <title>CSS API on style element in non-parser-created document</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <link rel=help href="https://bugzilla.mozilla.org/show_bug.cgi?id=1991800">
  </head>
  <body>
    <script>
      test(function() {
        let doc = document.implementation.createHTMLDocument("");
        let style = doc.createElement("style");
        doc.head.appendChild(style)
        assert_true(!!style.sheet, "The style element should have the sheet property.");
      }, "CSS API should work on non-parser-created documents");
    </script>
  </body>
</html>
